WMR: Don't use serializable transaction isolation
Seen in https://github.com/MaterializeInc/database-issues/issues/9439

My understanding is that `t1` was missing when running the query, thus
it failed, causing RQG to recreate the table, thus causing the apparent
wrong results in the other running threads. In the end this seems to be
badly designed in RQG, but I'm not that familiar with the design to fix
it in another way.